refactor(validation): address Copilot review on PR #82
- Split `in_rule.dart` into `in.dart` and `in_list.dart` to match the
  one-rule-per-file convention used across `validation/rules/` (required.dart,
  email.dart, unique.dart, etc.).
- Wrap caller-provided whitelist in `List.unmodifiable()` inside both `In` and
  `InList` constructors — external mutation can no longer drift validation
  behavior or the `:values` params after the rule is built.
