+++ lib/magic.dart
@@ -162,7 +162,8 @@ export 'src/validation/rules/confirmed.dart';
 export 'src/validation/rules/same.dart';
 export 'src/validation/rules/accepted.dart';
 export 'src/validation/rules/unique.dart';
-export 'src/validation/rules/in_rule.dart';
+export 'src/validation/rules/in.dart';
+export 'src/validation/rules/in_list.dart';
 export 'src/validation/exceptions/validation_exception.dart';
 export 'src/validation/exceptions/authorization_exception.dart';
 export 'src/validation/validator.dart';
