+++ lib/src/validation/rules/in_list.dart
@@ -1,35 +1,5 @@
 import '../contracts/rule.dart';
 
-/// Validates that the value is contained in a given whitelist.
-///
-/// Mirrors Laravel's `in:foo,bar,baz` rule. Null values pass so you can combine
-/// with [Required] to enforce presence separately.
-///
-/// ```dart
-/// Validator.make({'visibility': 'public'}, {
-///   'visibility': [In<String>(['public', 'private'])],
-/// });
-/// ```
-class In<T> extends Rule {
-  /// The allowed values.
-  final List<T> values;
-
-  In(this.values);
-
-  @override
-  bool passes(String attribute, dynamic value, Map<String, dynamic> data) {
-    if (value == null) return true;
-    if (value is! T) return false;
-    return values.contains(value);
-  }
-
-  @override
-  String message() => 'validation.in';
-
-  @override
-  Map<String, dynamic> params() => {'values': values.join(', ')};
-}
-
 /// Validates that the value matches an entry from an enum list.
 ///
 /// Accepts either the enum instance itself or a wire string. By default the
@@ -45,7 +15,7 @@ class In<T> extends Rule {
 /// });
 /// ```
 class InList<T extends Enum> extends Rule {
-  /// The allowed enum values.
+  /// The allowed enum values (immutable snapshot of the caller-provided list).
   final List<T> values;
 
   /// Whether string comparison ignores case.
@@ -54,7 +24,8 @@ class InList<T extends Enum> extends Rule {
   /// Optional mapper to translate an enum to its wire representation.
   final String Function(T value)? wire;
 
-  InList(this.values, {this.caseInsensitive = false, this.wire});
+  InList(List<T> values, {this.caseInsensitive = false, this.wire})
+    : values = List.unmodifiable(values);
 
   @override
   bool passes(String attribute, dynamic value, Map<String, dynamic> data) {
